Non-Debugged Editon of an Agent-Centric System.
PolicyChange and AVChange checkbox are not implemented yet, but are intended to/not to reset the agents learner on a Policy or AV Change, It also needs to be coded in that a World change always resets the learner.
(its a week bit screwy when you suddenly change the learners available actions and stuff).